+++ src/main/java/online/kingdomkeys/kingdomkeys/config/ModConfigs.java
@@ -54,6 +54,8 @@ public class ModConfigs {
 
     public static List<? extends Integer> hiddenMagic;
     public static boolean cmHeaderTextVisible, cmClassicColors, snapChatToCommandMenu, hpShowHearts, summonTogether, auto3rdPersonShip, cmChangeColor, customFont, shoulderSurfingDecoupled, seasonalEvents, portrait3D;
+
+    public static boolean bakeDisplayedModels, bakeHeldModels;
     public static int cmTextXOffset, cmSelectedXOffset, cmSubXOffset, hpAlarm, lockOnIconScale, lockOnIconRotation, lockOnHpPerBar, partyYDistance, cmEndLWidth, cmEndRWidth, cmHeaderEndLWidth, cmHeaderEndRWidth, cmReactionEndLWidth, cmReactionEndRWidth;
 
     public static void setHUDData(String name, List<? extends Float> data){
@@ -314,6 +316,9 @@ public class ModConfigs {
         cmSubXOffset = CLIENT.cmSubXOffset.get();
 
         portrait3D = CLIENT.portrait3D.get();
+
+        bakeDisplayedModels = CLIENT.bakeDisplayedModels.get();
+        bakeHeldModels = CLIENT.bakeHeldModels.get();
 
         hpShowHearts = CLIENT.hpShowHearts.get();
         hpAlarm = CLIENT.hpAlarm.get();